Fix CI compilation failures after refactoring
- Resolve 'ColFavourite' not declared error in 'sourcemodel.cpp' which
was left over from previous refactoring by commenting out the dangling
else-if block.
- Resolve 'enumerated mismatch in conditional expression' warning/error
in 'mainwindow.cpp' by using static_cast<int> for the roles.

+++ src/mainwindow.cpp
@@ -5113,8 +5113,9 @@ void MainWindow::setFavouriteRank() {
     int initialRank = 1;
     QModelIndex sourceIndex = proxy->mapToSource(selectedRows.first());
 
-    int favRole = (idRole == SourceModel::IdRole) ? SourceModel::FavouriteRole
-                                                  : SessionModel::FavouriteRole;
+    int favRole = (idRole == SourceModel::IdRole)
+                      ? static_cast<int>(SourceModel::FavouriteRole)
+                      : static_cast<int>(SessionModel::FavouriteRole);
     QVariant rankVal = model->data(sourceIndex, favRole);
     if (rankVal.isValid())
       initialRank = rankVal.toInt();
